Feat: FetchController 클래스 구현
